Routes and Controllers v1

+++ database/migrations/2024_07_17_173690_create_recipe_ingredients_table.php
@@ -15,16 +15,17 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('recipe_id');
             $table->unsignedBigInteger('ingredient_id');
-            $table->decimal('quantity', 8, 2); // Cambiar según tus necesidades
+            $table->integer('quantity'); // Cambiado a integer
             $table->string('unit')->nullable();
             $table->timestamps();
-
+        
             // Foreign keys
             $table->foreign('recipe_id')->references('id')->on('recipes')->onDelete('cascade');
             $table->foreign('ingredient_id')->references('id')->on('ingredients')->onDelete('cascade');
-
+        
             // Unique constraint to prevent duplicate entries
             $table->unique(['recipe_id', 'ingredient_id']);
+
         });
     }
